Refactor more crates
