+++ crates/nu-command-crates/bits/Cargo.toml
@@ -4,7 +4,7 @@ description = "Nushell's built-in commands"
 # repository = "https://github.com/nushell/nushell/tree/main/crates/nu-command"
 edition = "2021"
 license = "MIT"
-name = "nu-command-crates-formats"
+name = "nu-command-crates-bits"
 version = "0.70.1"
 
 
@@ -25,25 +25,4 @@ nu-term-grid = { path = "../../nu-term-grid", version = "0.70.1"  }
 nu-utils = { path = "../../nu-utils", version = "0.70.1" }
 nu-ansi-term = "0.46.0"
 
-csv = "1.1.6"
-eml-parser = "0.1.0"
-indexmap = { version="1.7", features=["serde-1"] }
-ical = "0.7.0"
-
-calamine = "0.18.0"
-toml = "0.5.8"
-roxmltree = "0.14.0"
-serde = { version="1.0.123", features=["derive"] }
-serde_urlencoded = "0.7.0"
-serde_ini = "0.2.0"
-serde_yaml = "0.9.4"
-rust-embed = "6.3.0"
-htmlescape = "0.3.1"
-quick-xml = "0.23.0"
-itertools = "0.10.0"
-fancy-regex = "0.10.0"
-chrono-humanize = "0.2.1"
-
-[target.'cfg(unix)'.dependencies]
-umask = "2.0.0"
-users = "0.11.0"
+num-traits = "0.2.14"
