+++ crates/nu-command-crates/bits/src/lib.rs
