+++ crates/nu-command-crates/bytes/src/bytes_.rs
@@ -36,14 +36,14 @@ impl Command for Bytes {
     }
 }
 
-#[cfg(test)]
-mod test {
-    use crate::Bytes;
+// #[cfg(test)]
+// mod test {
+//     use crate::Bytes;
 
-    #[test]
-    fn test_examples() {
-        use crate::test_examples;
+//     #[test]
+//     fn test_examples() {
+//         use crate::test_examples;
 
-        test_examples(Bytes {})
-    }
-}
+//         test_examples(Bytes {})
+//     }
+// }
