+++ crates/nu-command-crates/bytes/src/index_of.rs
@@ -1,4 +1,4 @@
-use crate::input_handler::{operate, CmdArgument};
+use nu_command_crates_input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::ast::{Call, CellPath};
 use nu_protocol::engine::{Command, EngineState, Stack};
@@ -211,14 +211,14 @@ fn search_all_index(input: &[u8], pattern: &[u8], from_end: bool, span: Span) ->
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
 
-    #[test]
-    fn test_examples() {
-        use crate::test_examples;
+//     #[test]
+//     fn test_examples() {
+//         use crate::test_examples;
 
-        test_examples(BytesIndexOf {})
-    }
-}
+//         test_examples(BytesIndexOf {})
+//     }
+// }
