+++ crates/nu-command-crates/bytes/src/remove.rs
@@ -1,4 +1,4 @@
-use crate::input_handler::{operate, CmdArgument};
+use nu_command_crates_input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::{
     ast::{Call, CellPath},
@@ -197,14 +197,14 @@ fn remove_impl(input: &[u8], arg: &Arguments, span: Span) -> Value {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
 
-    #[test]
-    fn test_examples() {
-        use crate::test_examples;
+//     #[test]
+//     fn test_examples() {
+//         use crate::test_examples;
 
-        test_examples(BytesRemove {})
-    }
-}
+//         test_examples(BytesRemove {})
+//     }
+// }
