+++ crates/nu-command-crates/bytes/src/replace.rs
@@ -1,4 +1,4 @@
-use crate::input_handler::{operate, CmdArgument};
+use nu_command_crates_input_handler::{operate, CmdArgument};
 use nu_engine::CallExt;
 use nu_protocol::{
     ast::{Call, CellPath},
@@ -172,14 +172,14 @@ fn replace_impl(input: &[u8], arg: &Arguments, span: Span) -> Value {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
 
-    #[test]
-    fn test_examples() {
-        use crate::test_examples;
+//     #[test]
+//     fn test_examples() {
+//         use crate::test_examples;
 
-        test_examples(BytesReplace {})
-    }
-}
+//         test_examples(BytesReplace {})
+//     }
+// }
