+++ crates/nu-command-crates/bytes/src/reverse.rs
@@ -1,4 +1,4 @@
-use crate::input_handler::{operate, CellPathOnlyArgs};
+use nu_command_crates_input_handler::{operate, CellPathOnlyArgs};
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
 use nu_protocol::ast::CellPath;
@@ -92,14 +92,14 @@ fn reverse(val: &Value, _args: &CellPathOnlyArgs, span: Span) -> Value {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
 
-    #[test]
-    fn test_examples() {
-        use crate::test_examples;
+//     #[test]
+//     fn test_examples() {
+//         use crate::test_examples;
 
-        test_examples(BytesReverse {})
-    }
-}
+//         test_examples(BytesReverse {})
+//     }
+// }
