+++ crates/nu-command-crates/deprecated/src/deprecated_commands.rs
