+++ crates/nu-command-crates/deprecated/src/hash_base64.rs
