+++ crates/nu-command-crates/deprecated/src/str_find_replace.rs
