+++ crates/nu-command-crates/deprecated/src/str_int.rs
