+++ crates/nu-command-crates/formats/src/to/csv.rs
@@ -95,7 +95,7 @@ fn to_csv(
 
 #[cfg(test)]
 mod test {
-    use super::*;
+    // use super::*;
 
     // #[test]
     // fn test_examples() {
