+++ crates/nu-command/Cargo.toml
@@ -33,6 +33,9 @@ nu-command-crates-shell = {path = "../nu-command-crates/shells", version = "0.70
 nu-command-crates-platform = {path = "../nu-command-crates/platform", version = "0.70.1"}
 nu-command-crates-formats = {path = "../nu-command-crates/formats", version = "0.70.1"}
 nu-command-crates-filesystem = {path = "../nu-command-crates/filesystem", version = "0.70.1"}
+nu-command-crates-deprecated = {path = "../nu-command-crates/deprecated", version = "0.70.1"}
+nu-command-crates-bits = {path = "../nu-command-crates/bits", version = "0.70.1"}
+nu-command-crates-bytes = {path = "../nu-command-crates/bytes", version = "0.70.1"}
 
 alphanumeric-sort = "1.4.4"
 base64 = "0.13.0"
@@ -67,7 +70,6 @@ meval = "0.2.0"
 mime = "0.3.16"
 notify = "4.0.17"
 num = { version = "0.4.0", optional = true }
-num-traits = "0.2.14"
 pathdiff = "0.2.1"
 powierza-coefficient = "1.0.1"
 quick-xml = "0.23.0"
@@ -100,10 +102,6 @@ sqlparser = { version = "0.23.0", features = ["serde"], optional = true }
 
 [target.'cfg(windows)'.dependencies]
 winreg = "0.10.1"
-
-[target.'cfg(unix)'.dependencies]
-umask = "2.0.0"
-users = "0.11.0"
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies.trash]
 version = "2.1.3"
