+++ crates/nu-command/src/lib.rs
@@ -1,11 +1,11 @@
-mod bits;
-mod bytes;
+// mod bits;
+// mod bytes;
 mod charting;
 mod conversions;
 mod core_commands;
 mod date;
 mod default_context;
-mod deprecated;
+// mod deprecated;
 mod env;
 mod example_test;
 mod experimental;
@@ -27,14 +27,14 @@ mod strings;
 mod system;
 mod viewers;
 
-pub use bits::*;
-pub use bytes::*;
+// pub use bits::*;
+// pub use bytes::*;
 pub use charting::*;
 pub use conversions::*;
 pub use core_commands::*;
 pub use date::*;
 pub use default_context::*;
-pub use deprecated::*;
+// pub use deprecated::*;
 pub use env::*;
 #[cfg(test)]
 pub use example_test::test_examples;
@@ -47,6 +47,9 @@ pub use hash::*;
 pub use math::*;
 pub use misc::*;
 // pub use network::*;
+pub use nu_command_crates_bits::*;
+pub use nu_command_crates_bytes::*;
+pub use nu_command_crates_deprecated::*;
 pub use nu_command_crates_filesystem::*;
 pub use nu_command_crates_network::*;
 pub use nu_command_crates_platform::*;
